Clang decides that 64-bit multiplication is the best way to extend a byte.  Revert to 64-bit algorithm for memset and memchr.

+++ src/string/memchr.c
@@ -5,21 +5,21 @@ void* memchr(const void* source, int c, size_t length)
 {
     const unsigned char* src = source;
 
-    for (; length && (uintptr_t) src % sizeof(unsigned long); --length)
+    for (; length && (uintptr_t) src % sizeof(uint64_t); --length)
         if (*src++ == c)
             return (void*)--src;
 
-    const unsigned long* vsrc = (const unsigned long*) src;
-    const unsigned long magic = (unsigned long) 0x7efefefefefefeffu;
-    const unsigned long mask = 0x0101010101010101u * (unsigned char) c;
+    const uint64_t* vsrc = (const uint64_t*) src;
+    const uint64_t magic = 0x7efefefefefefeffu;
+    const uint64_t mask = 0x0101010101010101u * (unsigned char) c;
 
-    for (; length >= sizeof(unsigned long); length -= sizeof(unsigned long))
+    for (; length >= sizeof(uint64_t); length -= sizeof(uint64_t))
     {
         src = (const unsigned char*) vsrc;
-        unsigned long word = *vsrc++ ^ mask;
+        uint64_t word = *vsrc++ ^ mask;
 
         if (((word + magic) ^ ~word) & ~magic)
-            for (int k = 0; k < sizeof(unsigned long); ++k)
+            for (int k = 0; k < sizeof(uint64_t); ++k)
                 if (src[k] == c)
                     return (void*)(src + k);
     }
